Very rough first cut
Screen scraping into a simple database is complete.  Zero work on the
UI and API have been done so far.  Next job is to get the crawler
running periodically and install onto the cloud so it can start
grabbing data in the background whilst the API and UI are being written.